+++ R/explain.R
@@ -1,6 +1,7 @@
 #' @export
 show_query.tbl_sql <- function(x, ...) {
-  message("<SQL>\n", remote_query(x))
+  cat_line("<SQL>")
+  cat_line(remote_query(x))
   invisible(x)
 }
 
@@ -14,10 +15,9 @@ show_query.tbl_lazy <- function(x, ...) {
 explain.tbl_sql <- function(x, ...) {
   force(x)
   show_query(x)
-
-  message("\n")
-  message("<PLAN>\n", remote_query_plan(x))
+  cat_line()
+  cat_line("<PLAN>")
+  cat_line(remote_query_plan(x))
 
   invisible(x)
 }
-