+++ tests/testthat/test-explain.R
@@ -1,6 +1,6 @@
-context("explain")
+context("test-explain.R")
 
 test_that("basic pipeline is correct", {
-  mf <- memdb_frame(x = 1:5) %>% filter(x > 5)
-  expect_output_file(explain(mf), "explain-sqlite.txt")
+  mf <- memdb_frame(x = 1:5, .name = "XYZ") %>% filter(x > 5)
+  expect_output_file(explain(mf), "test-explain-sqlite.txt")
 })